Physics system reworked

+++ src/Core/CameraFocusArea.h
@@ -6,7 +6,7 @@
 class CameraFocusArea : public Collider
 {
 public:
-    CameraFocusArea(const Vector2<int> &topLeft_, const Vector2<int> &size_);
+    CameraFocusArea(const Vector2<int> &topLeft_, const Vector2<uint16_t> &size_);
     Vector2<float> getCameraTargetPosition(const Vector2<float> &playerFocusPosition_);
     bool checkIfEnters(const Collider &cld_, bool isOwned_) const;
     float getScale() const;
@@ -14,7 +14,7 @@ public:
     void overrideFocusArea(const Collider &area_);
 
 private:
-    CameraFocusArea(const Vector2<int> &topLeft_, int scaledSizeX_, int scaledSizeY_);
+    CameraFocusArea(const Vector2<int> &topLeft_, uint16_t scaledSizeX_, uint16_t scaledSizeY_);
     
     Renderer &m_renderer;
     float m_scale;
